Build for android. Initial commit.

+++ libs/clipper/clipper.cpp
@@ -4568,7 +4568,7 @@ void OpenPathsFromPolyTree(PolyTree& polytree, Paths& paths)
 
 std::ostream& operator <<(std::ostream &s, const IntPoint &p)
 {
-  s << "(" << p.X << "," << p.Y << ")";
+  s << "(" << (long)(p.X) << "," << (long)(p.Y) << ")";
   return s;
 }
 //------------------------------------------------------------------------------
@@ -4578,8 +4578,8 @@ std::ostream& operator <<(std::ostream &s, const Path &p)
   if (p.empty()) return s;
   Path::size_type last = p.size() -1;
   for (Path::size_type i = 0; i < last; i++)
-    s << "(" << p[i].X << "," << p[i].Y << "), ";
-  s << "(" << p[last].X << "," << p[last].Y << ")\n";
+    s << "(" << (long)(p[i].X) << "," << (long)(p[i].Y) << "), ";
+  s << "(" << (long)(p[last].X) << "," << (long)(p[last].Y) << ")\n";
   return s;
 }
 //------------------------------------------------------------------------------
